fix(rust-core): fix all 10 compile errors across all platforms
- tor_manager: import BridgeConfig from tor_guardmgr::bridge (public path)
  instead of private arti_client::config re-export (E0603)
- tor_manager: fix BridgeConfigBuilder From trait by using correct type (E0277)
- tor_manager: use tokio::io::split() for owned TcpStream splitting in
  SOCKS5 handler to satisfy 'static lifetime for tokio::spawn (E0597)
- dns: access TorEngine client via pub client() method instead of
  private field to fix E0616 and cascading E0282 type inference errors
- dns: add explicit DataStream type annotation on stream variable
- tunnel/macos: change TUN_MTU from u32 to u16 to match tun 0.8
  Configuration::mtu() signature (E0308)
- tunnel/linux: same TUN_MTU u32 → u16 fix (E0308)
- Cargo.toml: add tor-guardmgr = 0.22 with bridge-client feature

+++ rust-core/src/tunnel/linux.rs
@@ -18,7 +18,8 @@ use tokio::sync::oneshot;
 
 const TUN_NAME: &str = "nodex0";
 const TUN_ADDR: &str = "10.66.0.1";
-const TUN_MTU:  u32  = 1500;
+// FIX: tun 0.8 Configuration::mtu() expects u16, not u32.
+const TUN_MTU:  u16  = 1500;
 const RT_TABLE: &str = "200";
 
 static STOP_TX: OnceCell<Mutex<Option<oneshot::Sender<()>>>> = OnceCell::new();
